Refactor permissions and make tournament object fully cacheable (#3341)

+++ app/features/lfg/actions/lfg.server.ts
@@ -1,5 +1,6 @@
 import type { ActionFunctionArgs } from "react-router";
 import { requireUser } from "~/features/auth/core/user.server";
+import { requirePermission } from "~/modules/permissions/guards.server";
 import { errorToastIfFalsy, parseRequestPayload } from "~/utils/remix.server";
 import * as LFGRepository from "../LFGRepository.server";
 import { lfgActionSchema } from "../lfg-schemas";
@@ -14,17 +15,15 @@ export const action = async ({ request }: ActionFunctionArgs) => {
 	const posts = await LFGRepository.findAllPosts(user);
 	const post = posts.find((post) => post.id === data.id);
 	errorToastIfFalsy(post, "Post not found");
-	errorToastIfFalsy(
-		post.author.id === user.id || user.roles.includes("ADMIN"),
-		"Not your own post",
-	);
 
 	switch (data._action) {
 		case "DELETE_POST": {
+			requirePermission(post, "DELETE");
 			await LFGRepository.deletePost(data.id);
 			break;
 		}
 		case "BUMP_POST": {
+			requirePermission(post, "EDIT");
 			await LFGRepository.bumpPost(data.id);
 			break;
 		}
